+++ tests/parameters/test_process_constraints.py
@@ -1,11 +1,8 @@
 """Test the pc processing."""
-from copy import deepcopy
-
 import numpy as np
 import pandas as pd
 import pytest
 from estimagic.exceptions import InvalidConstraintError
-from estimagic.optimization.optimize import maximize
 from estimagic.parameters.constraint_tools import check_constraints
 from estimagic.parameters.process_constraints import (
     _replace_pairwise_equality_by_equality,
@@ -33,38 +30,6 @@ def test_empty_constraints_work():
     constraints = [{"query": "bla == 'blubb'", "type": "equality"}]
 
     check_constraints(params, constraints)
-
-
-def test_bug_from_copenhagen_presentation():
-    # make sure maximum of work hours is optimal
-    def u(params):
-        return params["work"]["hours"] ** 2
-
-    start_params = {
-        "work": {"hourly_wage": 25.5, "hours": 2_000},
-        "time_budget": 24 * 7 * 365,
-    }
-
-    def return_all_but_working_hours(params):
-        out = deepcopy(params)
-        del out["work"]["hours"]
-        return out
-
-    res = maximize(
-        criterion=u,
-        params=start_params,
-        algorithm="scipy_lbfgsb",
-        constraints=[
-            {"selector": return_all_but_working_hours, "type": "fixed"},
-            {
-                "selector": lambda p: [p["work"]["hours"], p["time_budget"]],
-                "type": "increasing",
-            },
-        ],
-        lower_bounds={"work": {"hours": 0}},
-    )
-
-    assert np.allclose(res.params["work"]["hours"], start_params["time_budget"])
 
 
 def test_to_many_bounds_in_increasing_constraint_raise_good_error():
